research/requirements.txt
# Research Environment — Requirements

# Core dependencies
pyyaml>=6.0
numpy>=1.24.0
requests>=2.31.0

# Optional: Cloud LLM support
groq>=0.4.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
